[Refactor] 패키지 구조 및 클래스 위치 변경 (#40)
