+++ Module-API/src/main/java/depromeet/api/domain/auth/dto/request/KakaoAuthRequest.java
@@ -4,7 +4,7 @@ package depromeet.api.domain.auth.dto.request;
 import lombok.Data;
 
 @Data
-public class AuthRequest {
+public class KakaoAuthRequest {
     private String idToken;
 
     private String accessToken;
