+++ Module-API/src/main/java/depromeet/api/domain/auth/dto/response/KakaoAuthResponse.java
@@ -1,0 +1,3 @@
+package depromeet.api.domain.auth.dto.response;
+
+public class KakaoAuthResponse {}
